Feature/rm last read provider (#36)

+++ src/integration-test/kotlin/integration/chainadapter/ChainAdapterIntegrationTest.kt
@@ -68,7 +68,7 @@ class ChainAdapterIntegrationTest {
                     adapter.getLastReadBlock(),
                     adapter.lastReadBlockProvider.getLastBlockHeight()
                 )
-                assertEquals(consumedBlocks.last(), adapter.getLastReadBlock())
+                assertEquals(consumedBlocks.last(), adapter.getLastReadBlock().toLong())
             }
         }
     }
@@ -149,8 +149,11 @@ class ChainAdapterIntegrationTest {
                 logger.info { consumedBlocks }
                 assertEquals(transactions, consumedBlocks.size)
                 assertEquals(consumedBlocks.sorted(), consumedBlocks)
-                assertEquals(adapter.getLastReadBlock(), adapter.lastReadBlockProvider.getLastBlockHeight())
-                assertEquals(consumedBlocks.last(), adapter.getLastReadBlock())
+                assertEquals(
+                    adapter.getLastReadBlock(),
+                    adapter.lastReadBlockProvider.getLastBlockHeight()
+                )
+                assertEquals(consumedBlocks.last(), adapter.getLastReadBlock().toLong())
             }
         }
     }
